Encapsulate bottomsheet use in WalletTextInput
