+++ mobile-app/app/screens/AppNavigator/screens/Loans/components/AddOrRemoveCollateralForm.tsx
@@ -4,13 +4,12 @@ import { InputHelperText } from '@components/InputHelperText'
 import { SymbolIcon } from '@components/SymbolIcon'
 import { ThemedIcon, ThemedScrollView, ThemedText, ThemedView } from '@components/themed'
 import { WalletTextInput } from '@components/WalletTextInput'
-import { useBottomSheetInternal } from '@gorhom/bottom-sheet'
 import { StackScreenProps } from '@react-navigation/stack'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
 import BigNumber from 'bignumber.js'
-import React, { useCallback, useEffect, useState } from 'react'
-import { Platform, TouchableOpacity, View } from 'react-native'
+import React, { useEffect, useState } from 'react'
+import { TouchableOpacity, View } from 'react-native'
 import { TokenData } from '@defichain/whale-api-client/dist/api/tokens'
 import NumberFormat from 'react-number-format'
 import { useSelector } from 'react-redux'
@@ -53,23 +52,7 @@ export const AddOrRemoveCollateralForm = React.memo(({ route }: Props): JSX.Elem
   const [collateralValue, setCollateralValue] = useState<string>('')
   const isConversionRequired = isAdd && token.id === '0' ? new BigNumber(collateralValue).gt(DFIToken.amount) : false
   const [isValid, setIsValid] = useState(false)
-  const { shouldHandleKeyboardEvents } = useBottomSheetInternal()
-  const handleOnFocus = useCallback(
-    () => {
-      if (Platform.OS === 'ios') {
-        shouldHandleKeyboardEvents.value = true
-      }
-    },
-    [shouldHandleKeyboardEvents]
-  )
-  const handleOnBlur = useCallback(
-    () => {
-      if (Platform.OS === 'ios') {
-        shouldHandleKeyboardEvents.value = true
-      }
-    },
-    [shouldHandleKeyboardEvents]
-  )
+
   const validateInput = (input: string): void => {
     const formattedInput = new BigNumber(input)
     if (formattedInput.isGreaterThan(available) || formattedInput.isLessThanOrEqualTo(0) || formattedInput.isNaN()) {
@@ -142,8 +125,7 @@ export const AddOrRemoveCollateralForm = React.memo(({ route }: Props): JSX.Elem
         onClearButtonPress={() => setCollateralValue('')}
         placeholder={translate('components/AddOrRemoveCollateralForm', 'Enter an amount')}
         style={tailwind('h-9 w-10/12 flex-grow')}
-        onBlur={handleOnBlur}
-        onFocus={handleOnFocus}
+        hasBottomSheet
       />
       <InputHelperText
         label={`${translate('components/AddOrRemoveCollateralForm', isAdd ? 'Available' : 'Current')}: `}
